feat: 메인페이지 디자인, 2번쨰 섹션
